% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etch.R
\name{etch}
\alias{etch}
\title{Etch}
\usage{
etch(slate, message, ..., level = NULL, app_id = NULL)
}
\arguments{
\item{slate}{Slate object.}

\item{message}{Message to log.}

\item{...}{Named fields and values.}

\item{level}{Logging level.}

\item{app_id}{Application ID.}
}
\description{
Etch
}
