% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile.R
\name{peruse}
\alias{peruse}
\title{Peruse}
\usage{
peruse(slate, app_id = NULL, start = NULL, end = NULL)
}
\arguments{
\item{slate}{Slate object.}

\item{app_id}{App ID.}

\item{start}{Start time.}

\item{end}{End time.}
}
\description{
Peruse
}
