% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slate.R
\name{redis_slate}
\alias{redis_slate}
\title{Create a Redis Slate}
\usage{
redis_slate(path, config = redis_slate_config())
}
\arguments{
\item{path}{Unix domain socket.}

\item{config}{List of config options.}
}
\description{
Create a Redis Slate
}
