% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slate.R
\name{slate}
\alias{slate}
\title{Create a Slate}
\usage{
slate(path, config = slate_config())
}
\arguments{
\item{path}{Unix domain socket.}

\item{config}{List of config options.}
}
\description{
Create a Slate
}
