% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile.R
\name{stare}
\alias{stare}
\title{Stare}
\usage{
stare(slate, app_id = NULL, last_id = NULL)
}
\arguments{
\item{slate}{A slate object.}

\item{app_id}{Vector of app IDs to watch.}

\item{last_id}{Starting ID.}
}
\description{
Stare
}
