% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\name{study}
\alias{study}
\title{Study}
\usage{
study(slate, app_id = NULL, start = NULL, end = NULL)
}
\arguments{
\item{slate}{Slate object.}

\item{app_id}{App ID.}

\item{start}{Start time.}

\item{end}{End time.}
}
\description{
Study
}
