% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slate.R
\name{chisel}
\alias{chisel}
\title{Modify Default Logging Settings}
\usage{
chisel(slate, app_id = NULL, level = NULL)
}
\arguments{
\item{slate}{A slate object.}

\item{app_id}{App ID.}

\item{level}{Default level.}
}
\description{
Modify Default Logging Settings
}
