% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etch.R
\name{etch}
\alias{etch}
\title{Etch}
\usage{
etch(slate, message, app_id = NULL, level = NULL)
}
\arguments{
\item{slate}{Slate object.}

\item{message}{Message to log.}

\item{app_id}{Application ID.}

\item{level}{Logging level.}
}
\description{
Etch
}
