% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slate.R
\name{slate}
\alias{slate}
\title{Create a Slate}
\usage{
slate(path, default_app_id = NULL, default_level = NULL)
}
\arguments{
\item{path}{Unix domain socket.}

\item{app_id}{App ID.}
}
\description{
Create a Slate
}
