% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\name{stare}
\alias{stare}
\title{Stare}
\usage{
stare(slate, app_ids = NULL, id = NULL)
}
\arguments{
\item{slate}{A slate object.}

\item{app_ids}{Vector of app IDs to watch.}

\item{id}{Starting ID.}
}
\description{
Stare
}
