% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\name{study}
\alias{study}
\title{Study}
\usage{
study(slate, start = NULL, end = NULL, level = NULL)
}
\arguments{
\item{slate}{Slate object.}

\item{start}{Start time.}

\item{end}{End time.}

\item{level}{Logging level (not implemented).}
}
\description{
Study
}
